adicionado viewes para os controlers

+++ pessoal/gastos/src/main/java/com/example/curso/boot/web/controller/BillCollectorController.java
@@ -5,12 +5,16 @@ import com.example.curso.boot.services.BillCollectorService;
 import javassist.NotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
+import java.util.Objects;
 
 @Slf4j
 @Controller
@@ -32,9 +36,15 @@ public class BillCollectorController {
     }
 
     @PostMapping("/salvar")
-    public String salvar(BillCollector billCollector, ModelMap model) throws NotFoundException {
-        service.save(billCollector);
-        return listar(model);
+    public String salvar(BillCollector billCollector, RedirectAttributes attr) {
+
+        if (!Objects.isNull(service.findByName(billCollector.getName()))) {
+            attr.addFlashAttribute("fail", String.format("Cobrador '%s' ja existe.", billCollector.getName()));
+        } else {
+            service.save(billCollector);
+            attr.addFlashAttribute("success", "Cobrador cadastrado com sucesso.");
+        }
+        return "redirect:/collectors/listar";
     }
 
     @GetMapping("/editar/{id}")
@@ -45,16 +55,22 @@ public class BillCollectorController {
     }
 
     @PostMapping("/editar")
-    public String editar(BillCollector billCollector) {
+    public String editar(BillCollector billCollector, RedirectAttributes attr) {
+
+        attr.addFlashAttribute("success", "Cobrador editado com sucesso.");
         service.update(billCollector);
 
         return "redirect:/collectors/listar";
     }
 
     @GetMapping("/excluir/{id}")
-    public String excluir(@PathVariable("id") Long id, ModelMap model) throws NotFoundException {
-        service.delete(id);
+    public String excluir(@PathVariable("id") Long id, ModelMap model) {
+        if (Objects.isNull(service.findById(id))) {
+            model.addAttribute("fail", "Este cobrador não existe.");
+        } else {
+            service.delete(id);
+            model.addAttribute("success", "Cobrador excluido.");
+        }
         return listar(model);
     }
-
 }
